+++ init.go
@@ -1,14 +1,13 @@
 package task
 
 import (
-	"fmt"
 	"io"
 	"os"
 
 	"github.com/go-task/task/v3/errors"
 )
 
-const defaultTaskfile = `# https://taskfile.dev
+const DefaultTaskfile = `# https://taskfile.dev
 
 version: '3'
 
@@ -22,15 +21,17 @@ tasks:
     silent: true
 `
 
-// InitTaskfile Taskfile creates a new Taskfile at path
+const DefaultTaskFilename = "Taskfile.yml"
+
+// InitTaskfile creates a new Taskfile at path
 func InitTaskfile(w io.Writer, path string) error {
 	if _, err := os.Stat(path); err == nil {
 		return errors.TaskfileAlreadyExistsError{}
 	}
 
-	if err := os.WriteFile(path, []byte(defaultTaskfile), 0o644); err != nil {
+	if err := os.WriteFile(path, []byte(DefaultTaskfile), 0o644); err != nil {
 		return err
 	}
-	fmt.Fprintf(w, "%s created in the current directory\n", defaultTaskfile)
+
 	return nil
 }